Move secrets to their dedicated space
