+++ src/sdk/utils/secrets/get-secret.ts
@@ -1,5 +1,5 @@
-import { doGetSecret } from "@cre/sdk/testhelpers/do-get-secret";
-import { awaitAsyncSecret } from "@cre/sdk/testhelpers/await-async-secret";
+import { doGetSecret } from "@cre/sdk/utils/secrets/do-get-secret";
+import { awaitAsyncSecret } from "@cre/sdk/utils/secrets/await-async-secret";
 import { LazyPromise } from "@cre/sdk/utils/lazy-promise";
 
 export const getSecret = (id: string): Promise<any> => {
